backend/requirements.txt
django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.3.3
python-dotenv==1.2.2
